feat: Move delete account to profile
